reorganizada nas pastas, passei o deleteNode para utils, apaguei as funções do BST da utils
