+++ src/tree_utils.h
@@ -4,21 +4,21 @@
 #include <vector>
 #include <chrono>
 
-struct Node 
+struct Node
 {
     std::string word;
     std::vector<int> documentIds;
-    Node* parent;
-    Node* left;
-    Node* right;
-    int height;    // usado na AVL
-    int isRed;     // usado na RBT
+    Node *parent;
+    Node *left;
+    Node *right;
+    int height; // usado na AVL
+    int isRed;  // usado na RBT
 };
 
-struct BinaryTree 
+struct BinaryTree
 {
-    Node* root;
-    Node* NIL;  // usado na RBT (Opcional)
+    Node *root;
+    Node *NIL; // usado na RBT (Opcional)
 };
 
 struct InsertResult
@@ -37,27 +37,17 @@ struct SearchResult
     //<Possíveis outras variáveis>
 };
 
-
-
-// TODO: Adicionar Docstring
-BinaryTree* createBinaryTree();
-// TODO: Adicionar Docstring
-InsertResult insert(BinaryTree *tree, const std::string &word, int documentId);
-// TODO: Adicionar Docstring
-SearchResult search(BinaryTree *tree, const std::string &word);
-
-
 // TODO: Adicionar docstring
-void printIndex(BinaryTree* tree);
+void printIndex(BinaryTree *tree);
 // TODO: Adicionar docstring
-void printTree(BinaryTree* tree);
-
-
+void printTree(BinaryTree *tree);
 
 /**
  * @brief Cria um novo nó com a palavra fornecida.
- * 
+ *
  * @param x Palavra a ser armazenada no nó.
  * @return Node* Ponteiro para o novo nó criado.
  */
-Node* createNode(std::string x);
+Node *createNode(std::string x);
+
+void deleteNode(Node *node);
